server: enable large file support in database

+++ server/szurubooru/model/post.py
@@ -217,7 +217,7 @@ class Post(Base):
     # content description
     type = sa.Column("type", sa.Unicode(32), nullable=False)
     checksum = sa.Column("checksum", sa.Unicode(64), nullable=False)
-    file_size = sa.Column("file_size", sa.Integer)
+    file_size = sa.Column("file_size", sa.BigInteger)
     canvas_width = sa.Column("image_width", sa.Integer)
     canvas_height = sa.Column("image_height", sa.Integer)
     mime_type = sa.Column("mime-type", sa.Unicode(32), nullable=False)
